:hammer: refactor: rendering, singlePlay 분리
